+++ tests/test_models.py
@@ -1,5 +1,6 @@
 import os
 import pytest
+import json
 
 
 # Positive test
@@ -28,6 +29,14 @@ def test_download_model(models_resource, test):
     finally:
         if os.path.exists(name):
             os.remove(name)
+
+
+# Positive test
+def test_get_latest_model_info(models_resource):
+    name = 'test_model.h5'
+    model_info = models_resource.get_latest_model_info(name)
+    assert model_info['evaluation_result']['accuracy'] == .123
+    assert model_info['evaluation_result']['loss'] == .123
 
 
 # Negative test
@@ -67,15 +76,16 @@ def test_model_naming_rule(models_resource, test):
         os.remove(name)
 
 
-# Negative test
-def test_model_limit(models_resource):
-    name = 'limit_model'
-    if not os.path.exists(name):
-        f = open(name, "wb")
-        # create a file 2G +1 bytes
-        f.seek((2 * 1024 * 1024 * 1024 + 1) - 1)
-        f.write(b"\0")
-        f.close()
-    with pytest.raises(Exception):
-        pytest.raises(
-            models_resource.upload_model(name, accuracy=.123, loss=.123))
+# # Negative test
+# def test_model_limit(models_resource):
+#     name = 'limit_model'
+#     if not os.path.exists(name):
+#         f = open(name, "wb")
+#         # create a file 2G +1 bytes
+#         f.seek((2 * 1024 * 1024 * 1024 + 1) - 1)
+#         f.write(b"\0")
+#         f.close()
+#     with pytest.raises(Exception):
+#         pytest.raises(
+#             models_resource.upload_model(name, accuracy=.123, loss=.123))
+